Ensure session returns a UserInterface
zend-expressive-session does not serialize PHP objects directly, but
instead passes them to `json_encode()`; this is done to prevent PHP
de/serialization vulnerabilities.

As such, the logic within the `PhpSession` adapter needed to change: it
needed to pull the relevant data from the `UserInterface` implementation
(the username and role), and serialize that data; in turn, when pulling
the data from the session, it needs to create a `UserInterface`
instance.

For this latter, we use an anonymous class implementing the interface,
and populating it based on the session data.