Fix workflow for LoginHandler
As pointed out by @jonsa, the original workflow did not work, because it
required that any endpoint requiring authentication would respond to
`POST` requests. Since this is almost certainly not what is expected, I
needed to refactor how to handle requests where credentials are
submitted.

The new workflow has the handler respond to two different routes. The
first, a `GET` request, composes only the session middleware and
handler. The second, a `POST` request, composes the session middleware,
authentication middleware, and the handler. This allows successful
credentials to lead back to the handler, which can then redirect to the
originally requested page.